EHN: tidy up utils
all functions in misc.
