+++ hyperspy/misc/material.py
@@ -1,0 +1,84 @@
+import numpy as np
+
+from hyperspy.misc.eds.elements import elements as elements_db
+
+def weight_to_atomic(elements, weight_percent):
+    """Convert weight percent (wt%) to atomic percent (at.%).
+
+    Parameters
+    ----------
+    elements: list of str
+        A list of element abbreviations, e.g. ['Al','Zn']
+
+    weight_percent: list of float
+        The weight fractions (composition) of the sample.
+
+    Returns
+    -------
+    atomic_percent : list
+        Composition in atomic percent.
+
+    """
+    atomic_weights = np.array(
+        [elements_db[element]['A'] for element in elements])
+    atomic_percent = weight_percent / atomic_weights / (
+        weight_percent / atomic_weights).sum() * 100
+    return atomic_percent.tolist()
+
+
+def atomic_to_weight(elements, atomic_percent):
+    """Convert atomic percent to weight percent.
+
+    Parameters
+    ----------
+    elements: list of str
+        A list of element abbreviations, e.g. ['Al','Zn']
+
+    atomic_percent: list of float
+        The atomic fractions (composition) of the sample.
+
+    Returns
+    -------
+    weight_percent : composition in weight percent.
+
+    """
+
+    atomic_weights = np.array(
+        [elements_db[element]['A'] for element in elements])
+
+    weight_percent = atomic_percent * atomic_weights / (
+        atomic_percent * atomic_weights).sum() * 100
+    return weight_percent.tolist()
+
+
+def density_of_mixture_of_pure_elements(elements, weight_percent):
+    """Calculate the density a mixture of elements.
+
+    The density of the elements is retrieved from an internal database. The
+    calculation is only valid if there is no interaction between the
+    components.
+
+    Parameters
+    ----------
+    elements: list of str
+        A list of element symbols, e.g. ['Al', 'Zn']
+    weight_percent: list of float
+
+    Returns
+    -------
+    density: The density in g/cm3.
+
+    Examples
+    --------
+
+    Calculate the density of modern bronze given its weight percent:
+    >>> density_of_mixture_of_pure_elements(("Cu", "Sn"), (88, 12))
+    8.6903187973131466i
+
+    """
+    densities = np.array(
+        [elements_db[element]['density'] for element in elements])
+    density = (weight_percent / densities / 100).sum() ** -1
+    return density
+
+
